drop support of Go 1.12
