+++ .github/workflows/main.yml
@@ -12,36 +12,25 @@ jobs:
           - macos-latest
           - windows-latest
         go:
-          - '1.7'  # minimum version that macos-latest supports
-          - '1.10' # last version that doesn't support go modules
-          - '1.11' # first version that supports go modules
-          - '1.x'  # latest version
+          - '1.13' # minimum version
+          - 'oldstable'
+          - 'stable'
     runs-on: ${{ matrix.os }}
 
     steps:
       - name: Set up Go ${{ matrix.go }}
-        uses: actions/setup-go@v2
+        uses: actions/setup-go@v4
         with:
           go-version: ${{ matrix.go }}
       - run: go version
-      - name: Set up GOPATH
-        shell: bash
-        run: |
-          echo "GOPATH=${{ github.workspace }}" >> "$GITHUB_ENV"
-          echo "${{ github.workspace }}/bin" >> "$GITHUB_PATH"
 
-      - uses: actions/checkout@v2
-        with:
-          path: src/github.com/mattn/goveralls
-          ref: ${{ github.event.pull_request.head.sha }}
+      - uses: actions/checkout@v3
       - name: build
         run: |
           go get ./...
           go install .
-        working-directory: src/github.com/mattn/goveralls
       - name: test
         run: goveralls -service=github -parallel -flagname="Unit-${{ matrix.os }}-Go-${{ matrix.go }}"
-        working-directory: src/github.com/mattn/goveralls
         env:
           COVERALLS_TOKEN: ${{ github.token }}
           GIT_BRANCH: ${{ github.head_ref }}
@@ -51,10 +40,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Set up Go
-        uses: actions/setup-go@v2
-        with:
-          go-version: '1.x'
-      - uses: actions/checkout@v2
+        uses: actions/setup-go@v4
+      - uses: actions/checkout@v3
         with:
           ref: ${{ github.event.pull_request.head.sha }}
       - name: finish
