+++ go.mod
@@ -1,10 +1,8 @@
 module github.com/mattn/goveralls
 
-go 1.17
+go 1.12
 
 require (
 	golang.org/x/mod v0.4.2
 	golang.org/x/tools v0.1.1
 )
-
-require golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
