--- gocover_lt1.8.go
@@ -1,8 +1,0 @@
-//go:build !go1.8
-// +build !go1.8
-
-package main
-
-func findRootPackage(rootDirectory string) string {
-	return ""
-}
